Removing LocalChannel architecture
All
***
- Clean dependencies (not used)
- Fix memory
- Change to Guava where possible
- Fix various coding issues

Digest
******
- Fix SecureRandom for Windows (Sun provider) and Linux (urandom)
- Add Global configuration for SecureRandom
- Add support for TLS 1.2 using Native Netty Uber jar, but reverting to
 standard of JRE if not possible

Common
******
- Finalize GUID and IntegerUUID implementation (new global transfer uid
 and current localchannel uuid) (deprecation of UUID)
- Fix WaarpShutdownHook exit issue
- Add ThreadLocalRandom for Java 6

R66
***
- Fix Threads usage (roughly R66: 2xCLIENT + 2xRUNNER(LIMIT) + 1xSERVER,
HTTP/REST: 1xCLIENT)
- Remove LocalChannel and associated threads (LocalChannel tasks now are
 based on the map of requests and simulation of localChannels through
  method calls and using LocalChannelReference object)
- Fix memory (mainly KeepAlive and Startup/Validation)
- Add simple Junit (Selenium) on Monitor and Admin Web
- Add Junits on snmp and http (UI) and monitoring
- Test with just start and stop for every DB types (notice: Oracle Bad Driver
 since Java 6 differs on jdbc url)
- Add Junits for Java Tasks (AddDigest, AddUuid)
- Add Junits for SpooledDirectoryTransfer
- Extract XML definition (deduplication of codes with Proxy)
- OWASP security fixes
- checkParameters on REST/RESTV2/HTTP method
- Fix issue on Commander not found any request of transfer

R66 Proxy
*********
- Add simple Junit (Selenium) on Monitor and Admin Web
- Add Junits

GatewayKernel
*************
- Add simple Junit Rest
- Improve CPU usage (mainly StringBuilder)
- checkParameters on REST/HTTP method

FTP
***
- Change SSL to TLS in FTPS connection

GWFTP
*****
- Fix memory (mainly KeepAlive)
- Add Junits

+++ WaarpR66/src/main/java/org/waarp/openr66/client/SendThroughClient.java
@@ -106,13 +106,12 @@ import org.waarp.openr66.protocol.utils.R66Future;
  * <br>
  * <br>
  *
- *
  * @see TestSendThroughClient Class as example of usage in test part
  */
 public abstract class SendThroughClient extends AbstractTransfer {
   protected final NetworkTransaction networkTransaction;
   protected LocalChannelReference localChannelReference;
-  protected DbTaskRunner taskRunner = null;
+  protected DbTaskRunner taskRunner;
 
   /**
    * @param future
@@ -125,10 +124,10 @@ public abstract class SendThroughClient extends AbstractTransfer {
    * @param networkTransaction
    * @param id
    */
-  public SendThroughClient(R66Future future, String remoteHost, String filename,
-                           String rulename, String fileinfo, boolean isMD5,
-                           int blocksize, long id,
-                           NetworkTransaction networkTransaction) {
+  protected SendThroughClient(R66Future future, String remoteHost,
+                              String filename, String rulename, String fileinfo,
+                              boolean isMD5, int blocksize, long id,
+                              NetworkTransaction networkTransaction) {
     super(SendThroughClient.class, future, filename, rulename, fileinfo, isMD5,
           remoteHost, blocksize, id, null);
     this.networkTransaction = networkTransaction;
@@ -222,7 +221,6 @@ public abstract class SendThroughClient extends AbstractTransfer {
         } catch (final OpenR66ProtocolNotYetConnectionException e) {
           logger.debug("Not Yet Connected", e);
           exc = e;
-          continue;
         }
       }
       if (exc != null) {
@@ -246,12 +244,11 @@ public abstract class SendThroughClient extends AbstractTransfer {
       // now start the send from external data
       return true;
     } finally {
-      if (taskRunner != null) {
-        if (future.isFailed()) {
-          try {
-            taskRunner.delete();
-          } catch (final WaarpDatabaseException e) {
-          }
+      if (taskRunner != null && future.isFailed()) {
+        try {
+          taskRunner.delete();
+        } catch (final WaarpDatabaseException ignored) {
+          // nothing
         }
       }
     }
@@ -301,7 +298,8 @@ public abstract class SendThroughClient extends AbstractTransfer {
           ChannelUtils
               .writeAbstractLocalPacket(localChannelReference, validPacket,
                                         true);
-        } catch (final OpenR66ProtocolPacketException e) {
+        } catch (final OpenR66ProtocolPacketException ignored) {
+          // nothing
         }
         if (!localChannelReference.getFutureRequest().awaitOrInterruptible()) {
           // valid it however
@@ -309,18 +307,19 @@ public abstract class SendThroughClient extends AbstractTransfer {
               localChannelReference.getFutureEndTransfer().getResult());
         }
         if (taskRunner != null && taskRunner.isSelfRequested()) {
-          ChannelUtils.close(localChannelReference.getLocalChannel());
+          localChannelReference.close();
         }
       } else {
         transferInError(null);
       }
     } finally {
       if (taskRunner != null) {
-        if ((future.isDone() && (!future.isSuccess())) || nolog ||
+        if (future.isDone() && !future.isSuccess() || nolog ||
             taskRunner.shallIgnoreSave()) {
           try {
             taskRunner.delete();
-          } catch (final WaarpDatabaseException e) {
+          } catch (final WaarpDatabaseException ignored) {
+            // nothing
           }
         }
       }
@@ -347,11 +346,12 @@ public abstract class SendThroughClient extends AbstractTransfer {
       try {
         ChannelUtils
             .writeAbstractLocalPacket(localChannelReference, error, true);
-      } catch (final OpenR66ProtocolPacketException e1) {
+      } catch (final OpenR66ProtocolPacketException ignored) {
+        // nothing
       }
       localChannelReference.invalidateRequest(result);
     }
-    ChannelUtils.close(localChannelReference.getLocalChannel());
+    localChannelReference.close();
   }
 
   /**
